+++ hfny.jay/Stored Procedures/dbo.spAddcodeSubTopicDetail.sql
@@ -1,3 +1,4 @@
+
 SET QUOTED_IDENTIFIER ON
 GO
 SET ANSI_NULLS ON
@@ -6,7 +7,7 @@ CREATE PROCEDURE [dbo].[spAddcodeSubTopicDetail](@CompareToDate varchar(10)=NULL
 @DaysRequired int=NULL,
 @Interval varchar(40)=NULL,
 @RequiredBy varchar(4)=NULL,
-@SATFK int=NULL,
+@SATFK money=NULL,
 @SubTopicFK int=NULL,
 @SubTopicName varchar(64)=NULL,
 @TopicFK int=NULL,
