+++ hfny.jay/Stored Procedures/dbo.spEditcodeSAT.sql
@@ -1,8 +1,9 @@
+
 SET QUOTED_IDENTIFIER ON
 GO
 SET ANSI_NULLS ON
 GO
-CREATE PROCEDURE [dbo].[spEditcodeSAT](@codeSATPK int=NULL,
+CREATE PROCEDURE [dbo].[spEditcodeSAT](@codeSATPK money=NULL,
 @codeSATPK_old int=NULL,
 @ProgramFK int=NULL,
 @SATCompareDateField char(8)=NULL,
